third_party/WebKit/LayoutTests/editing/selection/mixed-editability-3.html

<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="./resources/js-test-selection-shared.js"></script>

<div contenteditable="true" id="test">
<span id="e1">1 editable</span>
<table border="1" contenteditable="false"><tr><td contenteditable="true">editable</td></tr></table>
<span id="e2">2 editable</span>
</div>

<script>
description('This tests moving the caret forward through content of mixed editability.  The caret should not be in non-editable piece.');

testCaretMotion({
    'direction': 'both',
    'granularity': 'line',
    'origin': 'e1',
    'target': 'e2',
});

testCaretMotion({
    'direction': 'both',
    'granularity': 'line',
    'origin': 'e1',
    'originOffset': -1,
    'target': 'e2',
    'targetOffset': -1,
});

if (window.testRunner)
    $('test').outerHTML = '';
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
