third_party/WebKit/LayoutTests/editing/selection/mixed-editability-3-expected.txt

This tests moving the caret forward through content of mixed editability. The caret should not be in non-editable piece.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Move forward line x 1
PASS selection.anchorNode == '[object Text]' is true
PASS selection.anchorOffset is 0
Move backward line x 1
PASS selection.anchorNode == '[object Text]' is true
PASS selection.anchorOffset is 0
Move forward line x 1
PASS selection.anchorNode == '[object Text]' is true
PASS selection.anchorOffset is 10
Move backward line x 1
PASS selection.anchorNode == '[object Text]' is true
PASS selection.anchorOffset is 10
PASS successfullyParsed is true

TEST COMPLETE
